feat: implement multi-step signup form with validation and payment processing
